Tích hơp phương thức thanh toán paypal

+++ src/services/OrderService.js
@@ -14,6 +14,8 @@ const createOrder = (newOrder) => {
             address,
             fullName,
             city,
+            isPaid,
+            paidAt,
             phone,
             user,
             deliveryMethod, // Get deliveryMethod from newOrder
@@ -78,6 +80,8 @@ const createOrder = (newOrder) => {
                 deliveryMethod, // Add deliveryMethod here
                 itemsPrice,
                 shippingPrice,
+                isPaid,
+                paidAt,
                 totalPrice,
                 user, // Add user to the order
             });
@@ -147,29 +151,58 @@ const getDetailsOrder = (id) => {
         }
     })
 }
-const cancelDetailsOrder = (id) => {
-    return new Promise(async (resolve, reject) => {
-        try {
-            const foundOrder = await Order.findOne({
-                _id: id
-            });
-            if (foundOrder === null) {
-                resolve({
-                    status: "ERR",
-                    message: "The order is not defined",
-                })
+const cancelDetailsOrder = async (id, data) => {
+    try {
+        // Xử lý cập nhật số lượng tồn kho cho các sản phẩm trong đơn hàng
+        const failedItems = [];
+        for (const order of data) {
+            const productData = await Product.findOneAndUpdate(
+                { _id: order.product },
+                {
+                    $inc: {
+                        countInStock: +order.amount, // Tăng số lượng tồn kho
+                        selled: -order.amount, // Giảm số lượng đã bán
+                    },
+                },
+                { new: true }
+            );
+
+            if (!productData) {
+                failedItems.push(order.product);
             }
-            await Order.findByIdAndDelete(id)
-            resolve({
-                status: "OK",
-                message: "DELETE order SUCCESS",
-                data: foundOrder
-            });
-        } catch (e) {
-            reject(e)
         }
-    })
-}
+
+        if (failedItems.length) {
+            return {
+                status: 'ERR',
+                message: `Các sản phẩm với ID ${failedItems.join(", ")} không tồn tại hoặc không thể cập nhật.`,
+            };
+        }
+
+        // Xóa đơn hàng sau khi tất cả các sản phẩm đã được cập nhật
+        const deletedOrder = await Order.findByIdAndDelete(id);
+
+        if (!deletedOrder) {
+            return {
+                status: 'ERR',
+                message: 'Xóa đơn hàng thất bại',
+            };
+        }
+
+        return {
+            status: 'OK',
+            message: 'Hủy đơn hàng thành công',
+            data: deletedOrder,
+        };
+    } catch (e) {
+        console.error("Lỗi khi hủy đơn hàng:", e.message);
+        return {
+            status: 'ERR',
+            message: e.message,
+        };
+    }
+};
+
 
 module.exports = {
     createOrder,
